Wire the dApp to the on-chain program
Buying cover is now a single instruction to the protocol program, which
creates the policy account and moves the premium into the program-owned
vault itself. Filing a claim calls file_claim, where the program checks
the signer owns the policy and that it is still claimable — the UI is no
longer trusted about either.

Reading policies is one filtered getProgramAccounts (discriminator plus
holder at offset 40) instead of replaying memo history and
reconstructing status from a stream of events. Status is a field the
program writes, so the client cannot disagree with the chain about it.

Two supporting fixes:
- Helius' free tier answers getProgramAccounts with a 503, so the worker
  proxy routes that one method to the public RPC, restricted to this
  program so it can't be used as a general indexer against our quota.
- policy ids are account addresses now, so the table and certificate
  abbreviate them rather than overflowing.

Verified the hand-rolled decoding against live devnet data: the filtered
query returns this holder's policy and every field round-trips —
flight, date, payout 10000, premium 240, status paid, oracle basis.

+++ components/PolicyCertificate.tsx
@@ -144,7 +144,8 @@ export default function PolicyCertificate({
 
           {/* data grid */}
           {[
-            { label: "CERTIFICATE No.", value: policy.id, color: "#22d3ee" },
+            // the policy's on-chain account address — abbreviated to fit
+            { label: "CERTIFICATE No.", value: short(policy.id), color: "#22d3ee" },
             { label: "POLICYHOLDER", value: short(holder), color: "#eae6ff" },
             { label: "FLIGHT", value: policy.flight, color: "#eae6ff" },
             { label: "COVERED DATE", value: policy.date, color: "#eae6ff" },
